bot: add btc/ltc/doge auto trading bot
This is new design bot that it won't make the order until it hit our
target price. It will monitor 3 coins at the same time and try to
make the order when the price hits our target price.
So, it won't waste the money by placing them on the orders, only places
orders when needed.
BTW, it doesn't check if the order we placed success or not, so we might
get some open orders. We have to cancel them manually.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>