bot: change the buy_count rule
We don't want to sell too much when price goes up if we have bought
more than 2 times. So, we discount the buy_count by 2 that will prevent bot
from selling them in low price.
BTW, we don't have to sleep when big drop is detected, it's kind of useless.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>