algo: acelan: adjust sell amount
Orginally, we will adjust the sell amount when buy_count larger or equal to 1.
In this case, buy one time and sell one time won't sell out enough amount of
coins, and we'll pile too much coin on hand.
So, change the rule a little to sell the same portion as we just bought when
buy_count equal to 1.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>