bot: move all coin related settings to config file
add 'base_coin' and 'target_coin' in config file, so that we can easily
change our base and target coin just in config file. These 2 variables
is necessary and could not be empty.
add 'min_target_coin' to set the buy back lower bound.

add my_bot.php, it's a more generic bot file, all we have to do is to
provide it an algo and a config file, then it should all work.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>