lib: add prev_orders to track our order numbers
It looks like sometimes cryptsy server replies the wrong order data,
that will lead to the algo to cancel all orders and place new orders.
It's annoying and will make the new price not following the rules.
So, try to add prev_orders variable to check again when the size of
orders is changed.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>